Materials to Prefabs + Cost Calculator
Turned the materials into prefabs so that they could hold a cost variable in addition to a material. Had to modify a lot of existing script to make the change, but it's mostly just changing types here and there.
With that done I made a new UI element to display the cost and a bit more script in the car controller to calculate said cost. Every car model and paint job has a different cost so the changes should be easy to spot.